requirements.txt
yt-dlp
discord.py[voice]
python-dotenv
orjson
niconico.py
httpx
tzdata
